+++ proj/src/PixelTransformer.hpp
@@ -8,7 +8,7 @@ class ViewConfig;
 class PixelTransformer {
  public:
   PixelTransformer(int width, int height, const ViewConfig& viewConfig);
-  Point3D transform(int x, int y) const;
+  Point3D transform(double x, double y) const;
  private:
   Matrix4x4 transformMatrix;
 };
